using more separate components
