+++ frontend/src/components/ProjectsComponents/InputFields/Coach/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Coach";
