+++ frontend/src/components/ProjectsComponents/InputFields/NumberOfStudents/index.ts
@@ -1,0 +1,1 @@
+export {default} from "./NumberOfStudents"
