+++ frontend/src/components/ProjectsComponents/InputFields/index.ts
@@ -1,0 +1,3 @@
+export { default as NameInput } from "./Name";
+export { default as NumberOfStudentsInput } from "./NumberOfStudents";
+export { default as CoachInput } from "./Coach";
