+++ frontend/src/components/ProjectsComponents/InputFields/styles.ts
@@ -1,0 +1,24 @@
+import styled from "styled-components";
+
+export const Input = styled.input`
+    margin-top: 10px;
+    padding: 5px 10px;
+    background-color: #131329;
+    color: white;
+    border: none;
+    border-radius: 5px;
+`;
+
+export const AddButton = styled.button`
+    padding: 5px 10px;
+    background-color: #00bfff;
+    color: white;
+    border: none;
+    margin-left: 5px;
+    border-radius: 5px;
+`;
+
+export const WarningContainer = styled.div`
+    max-width: fit-content;
+    margin-top: 10px;
+`;
