+++ frontend/src/views/projectViews/CreateProjectPage/CreateProjectPage.tsx
@@ -5,15 +5,18 @@ import {
     RemoveButton,
     CreateButton,
     AddedCoach,
-    WarningContainer,
 } from "./styles";
 import { createProject } from "../../../utils/api/projects";
 import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 import { GoBack } from "../ProjectDetailPage/styles";
 import { BiArrowBack } from "react-icons/bi";
-import { useNavigate } from "react-router-dom";
-import Alert from "react-bootstrap/Alert";
 import { TiDeleteOutline } from "react-icons/ti";
+import {
+    NameInput,
+    NumberOfStudentsInput,
+    CoachInput,
+} from "../../../components/ProjectsComponents/InputFields";
 
 export default function CreateProjectPage() {
     const [name, setName] = useState("");
@@ -23,9 +26,6 @@ export default function CreateProjectPage() {
 
     const [coach, setCoach] = useState("");
     const [coaches, setCoaches] = useState<string[]>([]);
-    const availableCoaches = ["coach1", "coach2", "admin1", "admin2"]; // TODO get users from API call
-
-    const [showAlert, setShowAlert] = useState(false);
 
     const navigate = useNavigate();
 
@@ -36,52 +36,18 @@ export default function CreateProjectPage() {
                 Cancel
             </GoBack>
             <h2>New Project</h2>
-            <Input
-                value={name}
-                onChange={e => setName(e.target.value)}
-                placeholder="Project name"
+            <NameInput name={name} setName={setName} />
+            <NumberOfStudentsInput
+                numberOfStudents={numberOfStudents}
+                setNumberOfStudents={setNumberOfStudents}
+            />
+            <CoachInput
+                coach={coach}
+                setCoach={setCoach}
+                coaches={coaches}
+                setCoaches={setCoaches}
             />
 
-            <div>
-                <Input
-                    type="number"
-                    min="0"
-                    value={numberOfStudents}
-                    onChange={e => {
-                        setNumberOfStudents(e.target.valueAsNumber);
-                    }}
-                    placeholder="Number of students"
-                />
-            </div>
-            <div>
-                <Input
-                    value={coach}
-                    onChange={e => setCoach(e.target.value)}
-                    list="users"
-                    placeholder="Coach"
-                />
-                <datalist id="users">
-                    {availableCoaches.map((availableCoach, _index) => {
-                        return <option key={_index} value={availableCoach} />;
-                    })}
-                </datalist>
-
-                <AddButton
-                    onClick={() => {
-                        if (availableCoaches.some(availableCoach => availableCoach === coach)) {
-                            const newCoaches = [...coaches];
-                            newCoaches.push(coach);
-                            setCoaches(newCoaches);
-                            setShowAlert(false);
-                        } else setShowAlert(true);
-                    }}
-                >
-                    Add coach
-                </AddButton>
-                <WarningContainer>
-                    <BadCoachAlert show={showAlert} setShow={setShowAlert}></BadCoachAlert>
-                </WarningContainer>
-            </div>
             <div>
                 {coaches.map((element, _index) => (
                     <AddedCoach key={_index}>
@@ -115,15 +81,4 @@ export default function CreateProjectPage() {
             </CreateButton>
         </CreateProjectContainer>
     );
-}
-
-function BadCoachAlert({ show, setShow }: { show: boolean; setShow: (state: boolean) => void }) {
-    if (show) {
-        return (
-            <Alert variant="warning" onClose={() => setShow(false)} dismissible>
-                Please choose an option from the list
-            </Alert>
-        );
-    }
-    return null;
 }
